module-6/opentelemetry-jaeger-distributed-tracing-demo/jaeger-tracing-python-service/requirements.txt
flask==1.1.2
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-mysql
mysql-connector==2.2.9
